Update image URL
I made this as a child, when I was first learning JS and programming in general. Then I published this to the internet when I first made a GitHub account (still no clue why I chose to do this to this day).

Since I changed my GitHub username, the URL that the image was stored at  using GitHub pages no longer exists. I've updated it to use my domain instead.

Was scrolling through my really old GitHub projects, and it's nice to see what I was up to man years ago.